+++ packages/docs/site/CHANGELOG.md
@@ -3,6 +3,7 @@
 - added examples for `@svizzle/ui`'s viewports
 - upgraded `d3-array`, `d3-geo`, `lamb`, `svelte`, `eslint`
 - adopted `eslint-plugin-svelte`, linted
+- `Link` -> `HyperLink`
 
 ## `@svizzle/site` v0.4.9